add article controller test

+++ src/controllers/articleController.js
@@ -1,13 +1,13 @@
 const { StatusCodes } = require('http-status-codes')
 const Article = require('../models/articleModel')
 const Likes = require('../models/likesModel')
-const Good = require('../models/goodModel')
 const Vote = require('../models/voteModel')
 const User = require('../models/userModel')
+const ArticleService = require('../services/articleService')
+const UserService = require('../services/userService')
 const Comment = require('../models/commentModel')
 const { AppError } = require('../utils/appError')
 const ObjectId = require('mongoose').Types.ObjectId;
-const jwt = require('jsonwebtoken');
 const lookupLikes = [{
     $lookup: {
         from: 'likes',
@@ -48,34 +48,6 @@ const lookupUpvote = [{
     }
 }
 ]
-
-// const lookupTag = [
-//     { $unwind: '$tags' },
-//     {
-//         $lookup: {
-//             from: 'tags',
-//             localField: "tags._id",
-//             foreignField: "_id",
-//             as: "tags.info"
-//         }
-//     },
-//     { $unwind: '$tags.info' },
-//     {
-//         $group: {
-//             _id: '$_id',
-//             root: { $mergeObjects: '$$ROOT' },
-//             tags: { $push: '$tags' }
-//         }
-//     },
-//     {
-//         $replaceRoot: {
-//             newRoot: {
-//                 $mergeObjects: ['$root', '$$ROOT']
-//             }
-//         }
-//     }
-// ]
-
 const lookupTag = [
     {
         $lookup: {
@@ -85,40 +57,6 @@ const lookupTag = [
             as: "tags"
         }
     },
-    // {
-    //     $match: {
-    //         'tags.name': 'add'
-    //     }
-    // },
-
-    // {
-    //     $lookup:
-    //     {
-    //         from: "tags",
-    //         let: { article_tags: "$tags" },
-    //         pipeline: [
-    //             {
-    //                 $match:
-    //                 {
-    //                     $expr:
-    //                     {
-    //                         $and:
-    //                             [
-    //                                 // { $eq: ["$name", "add"] },
-    //                                 { $in: ["$name", ["add"]] },
-    //                                 // { $in: ['cool']
-    //                                 // { $eq: ["$$article_tags", "$_id"] }
-    //                                 { $in: ["$_id", "$$article_tags"] }
-    //                             ]
-
-    //                     }
-    //                 }
-    //             }
-    //         ],
-    //         as: "tags"
-    //     }
-    // }
-
 ]
 
 
@@ -140,12 +78,10 @@ const articleProject = {
         'user': 1,
         'author': '$author.username',
         'description': 1,
-        // 'good': "$good.users",
         'votes': 1,
         'likes': "$like.users",
         'tags': 1,
         'likeCount': { $size: "$like.users" },
-        // 'goodCount': { $size: '$good.users' },
         'createdAt': 1,
         'updatedAt': 1,
     }
@@ -225,40 +161,6 @@ const getAggregateStageById = (_id) => {
 
 }
 module.exports = {
-    create: async (req, res, next) => {
-        try {
-            const { url, tags, description } = req.body
-            const user = await User.findById(req.decoded.user._id)
-            const findArticle = await Article.findOne({ url: url }).lean()
-            if (findArticle) {
-                throw new AppError('AppError', StatusCodes.CONFLICT, 'sorry! this URL already exists', true)
-            }
-            const mapped = tags.map(tag => { return { locked: true, _id: tag } })
-            const article = await Article.create({
-                url: url,
-                description: description,
-                tags: mapped,
-                user: user._id,
-            })
-            await Likes.create({
-                users: [],
-                article: article._id,
-            })
-            // const g = await Good.create({
-            //     users: [],
-            //     article: article._id,
-            // })
-            await Vote.create({
-                upvoteUsers: [],
-                donwvoteUsers: [],
-                article: article._id,
-            })
-            const populated = await Article.findById(article._id).populate('tags')
-            res.json(populated)
-        } catch (e) {
-            next(e)
-        }
-    },
     lists: async (req, res, next) => {
         try {
             // const user = req.decoded.user
@@ -321,6 +223,69 @@ module.exports = {
             next(e)
         }
     },
+    getRank: async (req, res, next) => {
+        try {
+            const likesRanking = await Article.aggregate([
+                ...lookupLikes,
+                {
+                    "$project": {
+                        "_id": 1,
+                        "url": 1,
+                        'count': { $size: "$like.users" },
+                    }
+                },
+                aggregateSort('count', -1),
+                aggregateLimit(3),
+            ])
+            const voteRanking = await Article.aggregate([
+                ...lookupUpvote,
+                {
+                    "$project": {
+                        "_id": 1,
+                        "url": 1,
+                        "count": { $subtract: [{ $size: '$votes.upvoteUsers' }, { $size: '$votes.downvoteUsers' }] },
+                        'up': { $size: '$votes.upvoteUsers' },
+                        'down': { $size: '$votes.downvoteUsers' },
+                    }
+                },
+                aggregateSort('count', -1),
+                aggregateLimit(3),
+            ])
+            res.json({ likesRanking: likesRanking, voteRanking: voteRanking })
+        } catch (e) {
+            next(e)
+        }
+    },
+    create: async (req, res, next) => {
+        try {
+            const { url, tags, description } = req.body
+            const user = await User.findById(req.decoded.user._id)
+            const findArticle = await Article.findOne({ url: url }).lean()
+            if (findArticle) {
+                throw new AppError('AppError', StatusCodes.CONFLICT, 'sorry! this URL already exists', true)
+            }
+            const mapped = tags.map(tag => { return { locked: true, _id: tag } })
+            const article = await Article.create({
+                url: url,
+                description: description,
+                tags: mapped,
+                user: user._id,
+            })
+            await Likes.create({
+                users: [],
+                article: article._id,
+            })
+            await Vote.create({
+                upvoteUsers: [],
+                donwvoteUsers: [],
+                article: article._id,
+            })
+            const populated = await Article.findById(article._id).populate('tags')
+            res.status(StatusCodes.CREATED).json(populated)
+        } catch (e) {
+            next(e)
+        }
+    },
     update: async (req, res, next) => {
         try {
             const { _id, url, tags, description } = req.body
@@ -355,6 +320,9 @@ module.exports = {
                 user: req.decoded.user._id
             })
 
+            if (!article) {
+                throw new AppError('AppError', StatusCodes.NO_CONTENT, 'there is no article on this request', true)
+            }
             await Likes.findOneAndRemove({
                 article: _id,
             })
@@ -367,39 +335,13 @@ module.exports = {
             next(e)
         }
     },
-    addTags: async (req, res, next) => {
-        try {
-            const user = await User.findById(req.decoded.user._id)
-            const { _id, tags } = req.body
-            const tag_ids = tags.map(tag => tag._id)
-            const article = await Article.findOneAndUpdate(
-                { _id: _id, user: user._id },
-                { $addToSet: { tags: tag_ids } },
-                { new: true })
-            res.json(article)
 
-        } catch (e) {
-            next(e)
-        }
-    },
-    updateTag: async (req, res, next) => {
-        try {
-            const user = await User.findById(req.decoded.user._id)
-            const { _id, tags } = req.body
-            const tag_ids = tags.map(tag => tag._id)
-            const article = await Article.findOneAndUpdate(
-                { _id: _id, user: user._id },
-                { tags: tag_ids },
-                { new: true }
-            )
-            res.json(article)
-        } catch (e) {
-            next(e)
-        }
-    },
     likes: async (req, res, next) => {
         try {
             const user = await User.findById(req.decoded.user._id)
+            if (!user) {
+                throw new AppError('AppError', StatusCodes.NO_CONTENT, 'user not found', true)
+            }
             const { _id } = req.body
             const isLike = await Likes.findOne({ article: _id, users: { $in: [user._id] } })
             const update = isLike ? { $pull: { users: user._id } } : { $addToSet: { users: user._id } }
@@ -411,79 +353,6 @@ module.exports = {
             const stage = getAggregateStageById(_id)
             const populated = await Article.aggregate(stage)
             res.json(populated[0])
-        } catch (e) {
-            next(e)
-        }
-    },
-    addComment: async (req, res, next) => {
-        try {
-            const user = await User.findById(req.decoded.user._id)
-            const { article_id, parent_id, comment } = req.body
-            const parent = await Comment.findById(parent_id).select('depth')
-            depth = parent ? parent.depth + 1 : 1
-
-            const createdComment = await Comment.create(
-                {
-                    article: article_id,
-                    comment: comment,
-                    user: user._id,
-                    parent: parent_id,
-                    children: [],
-                    depth: depth
-                }
-            )
-            const updatedParent = await Comment.findByIdAndUpdate(createdComment.parent, {
-                $push: {
-                    children: createdComment._id
-                }
-            })
-            const result = await Comment.find({ article: article_id, parent: null })
-            res.status(StatusCodes.OK).json(result)
-        } catch (e) {
-            next(e)
-        }
-    },
-    getComment: async (req, res, next) => {
-        try {
-            const { article } = req.query
-            const result = await Comment.find({ article: article, parent: null })
-            res.json(result)
-        } catch (e) {
-            next(e)
-        }
-    },
-    getRank: async (req, res, next) => {
-        try {
-            const likesRanking = await Article.aggregate([
-                ...lookupLikes,
-                // ...lookupUpvote,
-                {
-                    "$project": {
-                        "_id": 1,
-                        "url": 1,
-                        // 'like': 1,
-                        'count': { $size: "$like.users" },
-                    }
-                },
-                aggregateSort('count', -1),
-                aggregateLimit(3),
-            ])
-            const voteRanking = await Article.aggregate([
-                ...lookupUpvote,
-                {
-                    "$project": {
-                        "_id": 1,
-                        "url": 1,
-                        // "votes": 1,
-                        "count": { $subtract: [{ $size: '$votes.upvoteUsers' }, { $size: '$votes.downvoteUsers' }] },
-                        'up': { $size: '$votes.upvoteUsers' },
-                        'down': { $size: '$votes.downvoteUsers' },
-                    }
-                },
-                aggregateSort('count', -1),
-                aggregateLimit(3),
-            ])
-            res.json({ likesRanking: likesRanking, voteRanking: voteRanking })
         } catch (e) {
             next(e)
         }
